requirements.txt
picamera
RPi.GPIO
Adafruit-ADS1x15
Adafruit-MotorHAT
